Implement course attaching for tariffs and subscriptions

+++ app/Components/Tariff/Repository.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace App\Components\Tariff;
 
 use App\Models\Course;
-use App\Models\Enum\CourseStatus;
 use App\Models\Enum\TariffStatus;
 use App\Models\Tariff;
+use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
 
 /**
@@ -59,11 +59,13 @@ class Repository extends \App\Common\BaseComponentRepository
      */
     public function attachCourses(Tariff $tariff, iterable $courses): void
     {
+        $tariff->load('courses');
         foreach ($courses as $course) {
-            if ($course->status === CourseStatus::DISABLED) {
-                throw new Exceptions\CannotAttachDisabledCourse($course);
+            if ($tariff->courses->where('id', $course->id)->count()) {
+                continue;
             }
-            $tariff->courses()->attach($course->id);
+
+            $tariff->courses()->attach($course->id, ['created_at' => Carbon::now()]);
         }
     }
 
@@ -75,6 +77,10 @@ class Repository extends \App\Common\BaseComponentRepository
     public function detachCourses(Tariff $tariff, iterable $courses): void
     {
         foreach ($courses as $course) {
+            if ($tariff->courses->where('id', $course->id)->count() === 0) {
+                continue;
+            }
+
             $tariff->courses()->detach($course->id);
         }
     }
